Change counter datatype from int to BigInteger

+++ src/SortProfiler.java
@@ -1,8 +1,10 @@
+import java.math.BigInteger;
+
 /**
  * The {@code SortProfiler} interface defines methods for counting statements for sorting algorithms
  */
 public interface SortProfiler {
-    long getBubbleSortStatementCount(MedicalRecords[] medicalRecords);
-    long getInsertionSortStatementCount(MedicalRecords[] medicalRecords);
-    long getSelectionSortStatementCount(MedicalRecords[] medicalRecords);
+    BigInteger getBubbleSortStatementCount(MedicalRecords[] medicalRecords);
+    BigInteger getInsertionSortStatementCount(MedicalRecords[] medicalRecords);
+    BigInteger getSelectionSortStatementCount(MedicalRecords[] medicalRecords);
 }
